fix integer index in dsfit: fitdecaysin

+++ dsfit.py
@@ -220,7 +220,7 @@ def fitdecaysin(xdata,ydata,fitparams=None,domain=None,showfit=False,showstartfi
     if fitparams is None:    
         FFT=scipy.fft(fitdatay)
         fft_freqs=scipy.fftpack.fftfreq(len(fitdatay),fitdatax[1]-fitdatax[0])
-        max_ind=np.argmax(abs(FFT[4:len(fitdatay)/2.]))+4
+        max_ind=np.argmax(abs(FFT[4:int(len(fitdatay)/2)]))+4
         fft_val=FFT[max_ind]
         
         fitparams=[0,0,0,0,0]
